+++ internal/command/testdata/test/dynamic_source_conflicting_test_files/modules/other/main.tf
@@ -1,0 +1,7 @@
+resource "test_resource" "foo" {
+  value = "baz"
+}
+
+output "value" {
+  value = test_resource.foo.value
+}